+++ EduMatch.BusinessLogicLayer/Services/TutorCertificateService.cs
@@ -145,7 +145,7 @@ namespace EduMatch.BusinessLogicLayer.Services
 					CertificateUrl = certUrl,
 					CertificatePublicId = certPublicId,
 					CreatedAt = DateTime.UtcNow,
-					Verified = VerifyStatus.Pending,
+					Verified = (int)VerifyStatus.Pending,
 					RejectReason = null
 				};
 
@@ -195,7 +195,7 @@ namespace EduMatch.BusinessLogicLayer.Services
 					CertificateUrl = request.CertificateUrl,
 					CertificatePublicId = null, // No public ID for external URLs
 					CreatedAt = DateTime.UtcNow,
-					Verified = VerifyStatus.Pending,
+					Verified = (int)VerifyStatus.Pending,
 					RejectReason = null
 				};
 
@@ -268,7 +268,7 @@ namespace EduMatch.BusinessLogicLayer.Services
 				// Handle Verified status
 				if (request.Verified.HasValue)
 				{
-					existingEntity.Verified = request.Verified.Value;
+					existingEntity.Verified = (int)request.Verified.Value;
 					if (request.Verified.Value == VerifyStatus.Rejected)
 					{
 						if (string.IsNullOrWhiteSpace(request.RejectReason))
@@ -354,7 +354,7 @@ namespace EduMatch.BusinessLogicLayer.Services
 				// Handle Verified status
 				if (request.Verified.HasValue)
 				{
-					existingEntity.Verified = request.Verified.Value;
+					existingEntity.Verified = (int)request.Verified.Value;
 					if (request.Verified.Value == VerifyStatus.Rejected)
 					{
 						if (string.IsNullOrWhiteSpace(request.RejectReason))
